+++ core/src/net/oschina/j2cache/ClusterPolicyFactory.java
@@ -18,6 +18,8 @@ package net.oschina.j2cache;
 import net.oschina.j2cache.redis.RedisClient;
 import net.oschina.j2cache.redis.RedisPubSubClusterPolicy;
 
+import java.util.Properties;
+
 /**
  * 集群策略工厂
  * @author Winter Lau(javayou@gmail.com)
@@ -28,11 +30,12 @@ public class ClusterPolicyFactory {
      * 使用 Redis 订阅和发布机制，该方法只能调用一次
      * @param name  频道名称
      * @param redis Redis 客户端接口
+     * @param props 框架配置
      * @return 返回 Redis 集群策略的实例
      */
-    public final static ClusterPolicy redis(String name, RedisClient redis) {
+    public final static ClusterPolicy redis(String name, RedisClient redis, Properties props) {
         RedisPubSubClusterPolicy policy = new RedisPubSubClusterPolicy(name, redis);
-        policy.connect();
+        policy.connect(props);
         return policy;
     }
 
@@ -40,12 +43,29 @@ public class ClusterPolicyFactory {
      * 使用 JGroups 组播机制
      * @param name  组播识别码
      * @param configPath 配置文件的名称
+     * @param props 框架配置
      * @return 返回 JGroups 集群策略的实例
      */
-    public final static ClusterPolicy jgroups(String name, String configPath) {
+    public final static ClusterPolicy jgroups(String name, String configPath, Properties props) {
         JGroupsClusterPolicy policy = new JGroupsClusterPolicy(name, configPath);
-        policy.connect();
+        policy.connect(props);
         return policy;
+    }
+
+    /**
+     * 加载自定义的集群通知策略
+     * @param classname
+     * @param props
+     * @return
+     */
+    public final static ClusterPolicy custom(String classname, Properties props) {
+        try {
+            ClusterPolicy policy = (ClusterPolicy)Class.forName(classname).newInstance();
+            policy.connect(props);
+            return policy;
+        } catch (Exception e) {
+            throw new CacheException("Failed in load custom cluster policy. class = " + classname, e);
+        }
     }
 
 }