+++ core/src/net/oschina/j2cache/util/SerializationUtils.java
@@ -22,6 +22,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.Serializable;
+import java.util.Arrays;
 
 /**
  * 对象序列化工具包
@@ -70,7 +71,12 @@ public class SerializationUtils {
             return null;
         if(obj instanceof Number || obj instanceof String || obj instanceof Character || obj instanceof Boolean)
             return obj.toString().getBytes();
-        return g_serializer.serialize(obj);
+
+        byte[] bytes = g_serializer.serialize(obj);
+        byte[] results = new byte[bytes.length + 1];
+        results[0] = 0x00;
+        System.arraycopy(bytes, 0, results, 1, bytes.length);
+        return results;
     }
 
     /**
@@ -80,8 +86,11 @@ public class SerializationUtils {
      * @throws IOException io exception
      */
     public static Serializable deserialize(byte[] bytes) throws IOException {
-        if(bytes == null)
+        if(bytes == null || bytes.length == 0)
             return null;
-        return g_serializer.deserialize(bytes);
+        if(bytes[0] != 0x00)
+            return new String(Arrays.copyOfRange(bytes, 1, bytes.length));
+
+        return g_serializer.deserialize(Arrays.copyOfRange(bytes, 1, bytes.length));
     }
 }